+++ app/build.gradle.kts
@@ -64,6 +64,8 @@ android {
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
 
         buildConfigField("String", "BUILD_COMMIT_HASH", "\"${getGitCommitHash()}\"")
+        buildConfigField("String", "API_VERSION", "\"v~draft2\"")
+        buildConfigField("String", "STORE_URL", "\"fladdonstest.patrickgold.dev\"")
 
         ksp {
             arg("room.schemaLocation", "$projectDir/schemas")
