+++ app/src/main/kotlin/dev/patrickgold/florisboard/app/ext/AddonBox.kt
@@ -1,0 +1,92 @@
+package dev.patrickgold.florisboard.app.ext
+
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Shop
+import androidx.compose.material.icons.outlined.FileDownload
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.unit.dp
+import dev.patrickgold.florisboard.app.LocalNavController
+import dev.patrickgold.florisboard.app.Routes
+import dev.patrickgold.florisboard.lib.android.launchUrl
+import dev.patrickgold.florisboard.lib.compose.FlorisOutlinedBox
+import dev.patrickgold.florisboard.lib.compose.FlorisTextButton
+import dev.patrickgold.florisboard.lib.compose.defaultFlorisOutlinedBox
+import dev.patrickgold.florisboard.lib.compose.stringRes
+import dev.patrickgold.florisboard.lib.ext.Extension
+import dev.patrickgold.florisboard.lib.ext.generateUpdateUrl
+import org.florisboard.lib.kotlin.curlyFormat
+
+@Composable
+fun UpdateBox(extensionIndex: List<Extension>) {
+    val context = LocalContext.current
+    FlorisOutlinedBox(
+        modifier = Modifier.defaultFlorisOutlinedBox(),
+    ) {
+        Text(
+            modifier = Modifier.padding(start = 16.dp, top = 8.dp, end = 16.dp, bottom = 4.dp),
+            text = "Since this app does not have Internet permission, updates for installed extensions must be checked manually.",
+            style = MaterialTheme.typography.bodySmall,
+        )
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(horizontal = 6.dp),
+        ) {
+            FlorisTextButton(
+                onClick = {
+                    context.launchUrl(extensionIndex.generateUpdateUrl(version = "v~draft2", host = "fladdonstest.patrickgold.dev"))
+                },
+                icon = Icons.Outlined.FileDownload,
+                text = "Search for Updates"
+            )
+            Spacer(modifier = Modifier.weight(1f))
+        }
+    }
+}
+
+@Composable
+fun AddonManagementReferenceBox(
+    type: ExtensionListScreenType
+) {
+    val navController = LocalNavController.current
+
+    FlorisOutlinedBox(
+        modifier = Modifier.defaultFlorisOutlinedBox(),
+        title = "Managing {extensions}".curlyFormat(
+            "extensions" to type.let { stringRes(id = it.titleResId).lowercase() }
+        )
+    ) {
+        Text(
+            modifier = Modifier.padding(horizontal = 16.dp, vertical = 4.dp),
+            text = "All tasks related to importing, exporting, creating, customizing, and removing extensions can be handled through the centralized addon manager.",
+            style = MaterialTheme.typography.bodySmall,
+        )
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(horizontal = 6.dp),
+        ) {
+            Spacer(modifier = Modifier.weight(1f))
+            FlorisTextButton(
+                onClick = {
+                    val route = Routes.Ext.List(type, showUpdate = true)
+                    navController.navigate(
+                        route
+                    )
+                },
+                icon = Icons.Default.Shop,
+                text = "Go to {ext_home_title}".curlyFormat(
+                    "ext_home_title" to stringRes(type.titleResId),
+                ),
+            )
+        }
+    }
+}
